test: consolidate grid component-renderer ITs into one view
Move the component-cell and component-details regenerated-on-refreshItem
scenarios from the demo GridViewUsingComponents page/IT into the
dedicated ComponentColumnsPage/IT, then delete the demo pair. These
exercise component regeneration on update, distinct from the existing
add-columns-before/after-attach tests, so nothing is dropped.
